+++ src/test/java/org/cicirello/search/problems/LCSTests.java
@@ -312,318 +312,4 @@ public class LCSTests {
     assertTrue(minCost >= problem.minCost());
     assertTrue(maxValue <= problem.maxValue());
   }
-
-  @Test
-  public void testDiffSize() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 0.5, 0.5);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertTrue(minCost >= problem.minCost());
-    assertTrue(maxValue <= problem.maxValue());
-
-    problem = new LargestCommonSubgraph(n2, n1, 0.5, 0.5);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertTrue(minCost >= problem.minCost());
-    assertTrue(maxValue <= problem.maxValue());
-  }
-
-  @Test
-  public void testDiffSizeSeed() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 0.5, 0.5, 42);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertTrue(minCost >= problem.minCost());
-    assertTrue(maxValue <= problem.maxValue());
-
-    problem = new LargestCommonSubgraph(n2, n1, 0.5, 0.5, 42);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertTrue(minCost >= problem.minCost());
-    assertTrue(maxValue <= problem.maxValue());
-  }
-
-  @Test
-  public void testDiffSizeFullyConnected() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 1.0, 1.0);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n2 * (n2 - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n2 * (n2 - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-
-    problem = new LargestCommonSubgraph(n2, n1, 1.0, 1.0);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n2 * (n2 - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n2 * (n2 - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testDiffSizeFullyConnectedSeed() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 1.0, 1.0, 42);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n2 * (n2 - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n2 * (n2 - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-
-    problem = new LargestCommonSubgraph(n2, n1, 1.0, 1.0, 42);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n2 * (n2 - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n2 * (n2 - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testDiffSizeNoEdges() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 0.0, 0.0);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-
-    problem = new LargestCommonSubgraph(n2, n1, 0.0, 0.0);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testDiffSizeNoEdgesSeed() {
-    int n1 = 7;
-    int n2 = 4;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n1, n2, 0.0, 0.0, 42);
-    assertEquals(n1, problem.size());
-    Permutation p = new Permutation(n1, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-
-    problem = new LargestCommonSubgraph(n2, n1, 0.0, 0.0, 42);
-    assertEquals(n1, problem.size());
-    p = new Permutation(n1, 0);
-    minCost = Integer.MAX_VALUE;
-    maxCost = Integer.MIN_VALUE;
-    minValue = Integer.MAX_VALUE;
-    maxValue = Integer.MIN_VALUE;
-    iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
 }
